backend/vad/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
PyJWT
sqlalchemy>=1.4
aiomysql
alembic
pydantic
passlib[bcrypt]
python-multipart